Separate reading a site from collecting it
brand-analysis.ts was 1816 lines doing two jobs. Fetching a site is the
same work for anyone — it resolves, it defends against the same tricks,
it parses the same way. Interpreting it is not: a social autopilot wants
colours and fonts to make graphics, a lead product wants to know what
you do NOT do, because the strongest rule in its drafter is never to
promise what is not there.

The boundary was already in the file, at line 1298: `// --- LLM
Analysis ---`. I cut there.

The package takes ~1360 lines — entry resolution, the SSRF guard, CDN
block-page detection, the real-browser fallback, internal page
discovery, Shopify and WooCommerce catalogues, and the extraction of
metadata, logos, colours, fonts, socials, text and archetype. Its 40
tests came with it, because they covered exactly that half.

The app keeps 655 and becomes what it actually is: the reader.

Two silent regressions the cut created, and neither test would have
caught. Making the renderer injectable means its default is "no
browser", and loadPageHtml and fetchInternantPages both took it from the
default — Browserless would have stopped being used with nothing
failing, returning empty pages on JS sites while fetchInternalPages'
own comment promises the opposite. The real renderer is now passed
explicitly at all three sites.

brand-analysis re-exports the thirty collection symbols because fourteen
modules import them from that address. It is a bridge, not a home.

New: crawlSite() returns the raw material with no model call at all,
which could not exist here before because collection and reading were
one block.

6146 tests green, no new type errors.
